2.5.4

+++ out/DeviceDetailsPanel.js
@@ -183,8 +183,18 @@ class DeviceDetailsPanel {
                                 if (appname !== null) {
                                     vscode.window.showInformationMessage("Application '" + appname + "' will be deleted from the device", ...["Delete Application", "Cancel"]).then((answer) => {
                                         if (answer === "Delete Application") {
-                                            this.doDeleteApp(device.mac, appname);
-                                            Providers_1.Providers.btdevices.showWait(true);
+                                            if (appname.toLowerCase() == 'platform.cub') {
+                                                vscode.window.showWarningMessage("ATTENTION! Removing this application will render the device's operating system inoperable. Recovery will be possible only by reinstalling the firmware. Are you sure you want to continue?", ...["YES, Delete Application", "Cancel"]).then((answer) => {
+                                                    if (answer === "YES, Delete Application") {
+                                                        this.doDeleteApp(device.mac, appname);
+                                                        Providers_1.Providers.btdevices.showWait(true);
+                                                    }
+                                                });
+                                            }
+                                            else {
+                                                this.doDeleteApp(device.mac, appname);
+                                                Providers_1.Providers.btdevices.showWait(true);
+                                            }
                                         }
                                         else {
                                             this._panel.webview.postMessage({ type: 'endRequest' });
